fix-timer
# タイマー機能をGameManagerに移行した
- CountdownTimerを消去
- 代わりにGameManagerから残り時間を受け取って表示するTimerUIを追加
-　必要ないNavMesh関連のコンポーネントをデタッチ、PlayerPrefsKesys.csを消去
- 残り時間のデザインを少し変更
