+++ Assets/AddressableAssetsData/AddressableAssetSettings.asset
@@ -15,7 +15,7 @@ MonoBehaviour:
   m_DefaultGroup: 597962e2086834640b46eb6f7ac771ca
   m_currentHash:
     serializedVersion: 2
-    Hash: bb194e4fb162d0bfe87bb6d73589c656
+    Hash: 830d7c4d2a082a2ab2e1053f32193447
   m_OptimizeCatalogSize: 0
   m_BuildRemoteCatalog: 0
   m_CatalogRequestsTimeout: 0
