--- Assets/Sandbox/Uchiyama/Scenes/Global Volume Profile.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 750fa1d0d1dea4fd4a7744d701047365
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
