+++ Assets/Sandbox/Uchiyama/Scenes/MainScene_u.unity.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0bae7cc4ccf074fcdbc2d13b14eaeae0
+guid: 6bc95c0add4e049a3a410f43bde3613d
 DefaultImporter:
   externalObjects: {}
   userData:
