--- Assets/Sandbox/Uchiyama/Scenes/SpherePrototype_u.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d15b6f752a60a40b79927be924567bba
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
