--- Assets/Sandbox/Uchiyama/Scenes/SpherePrototype_u.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 38bca7325f0b546b8b2b5f4d2799b5c7
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
