--- Assets/Sandbox/Uchiyama/Scenes/SpherePrototype_u/NavMesh-Floor (1) 1.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e929dfb2b906544adbe640e8dfa52068
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 23800000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
