--- Assets/Sandbox/Uchiyama/Scenes/SpherePrototype_u/NavMesh-Floor (1) 2.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f6e19222592024dc6925099ab818587e
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 23800000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
