--- Assets/Sandbox/Uchiyama/Scenes/SpherePrototype_u/NavMesh-Floor (1) 3.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ed8634e4fadbf4f3bb065615410de5fb
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 23800000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
