--- Assets/Sandbox/Uchiyama/Scenes/SpherePrototype_u/NavMesh-Floor (1).asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c0695ee86cbff48b2a13ac8954125815
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 23800000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
