--- Assets/Sandbox/Uchiyama/Scenes/SpherePrototype_u/NavMesh-Floor 1.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 428ce41e1c98f47b3baaf94fb1d7c2b1
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 23800000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
