--- Assets/Sandbox/Uchiyama/Scenes/SpherePrototype_u/NavMesh-Floor 2.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b22290db3b4694015980bf45f8f90b98
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 23800000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
