--- Assets/Sandbox/Uchiyama/Scenes/SpherePrototype_u/NavMesh-Floor 3.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7a792172a6d4f4cb8987287b05bb1f25
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 23800000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
