--- Assets/Sandbox/Uchiyama/Scenes/SpherePrototype_u/NavMesh-Floor 4.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b70d2d6e953a649f7ba9ad2db8083deb
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 23800000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
