--- Assets/Sandbox/Uchiyama/Scenes/SpherePrototype_u/NavMesh-Floor 5.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 828bc26dc2d1c4b96970741961899918
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 23800000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
