--- Assets/Sandbox/Uchiyama/Scenes/SpherePrototype_u/NavMesh-Floor 6.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 64845f1d81b9544dabbda5bf8fce02dd
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 23800000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
