--- Assets/Sandbox/Uchiyama/Scenes/SpherePrototype_u/NavMesh-Floor.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 14a9636211f6348cb9a33966a55517ca
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 23800000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
