--- Assets/Sandbox/Uchiyama/Scenes/SpherePrototype_u/NavMesh-MainGameLevel.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: da56272a3487345cbaec3864dee8ccfd
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 23800000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
