--- Assets/Sandbox/Uchiyama/Scenes/SpherePrototype_u/NavMesh-NavMesh Surface.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5ba7ce1bdaefc4ff5a488b2ff0a2b67b
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 23800000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
