--- Assets/Sandbox/Uchiyama/Scenes/SpherePrototype_u/NavMesh-RevealableFloor (3).asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: dadd4315653d641e5863821dbe9a6eca
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 23800000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
