--- Assets/Scripts/System/CountdownTimer.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: b02ef6eec65b442748b19007077a90a2
