+++ Assets/Scripts/System/GameManager.cs.meta
@@ -1,2 +1,11 @@
 fileFormatVersion: 2
 guid: 1bfd954a057514525b652f928d677abb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
