--- Assets/Scripts/Util/PlayerPrefsKeys.cs
@@ -1,8 +1,0 @@
-// PlayerPrefsKeys.cs
-using UnityEngine;
-
-public static class PlayerPrefsKeys // staticクラスとして定義
-{
-    // 残り時間を保存するキー
-    // public const string RemainingTimeAtClear = "RemainingTimeAtClear"; 
-}
