--- Assets/Scripts/Util/PlayerPrefsKeys.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: b4e2ee8aeb89447fb9c6c73db6f0b76c
